Remove redundant casts

+++ src/debug/debug_gui.cpp
@@ -492,7 +492,7 @@ static void MakeSubWindows(void) {
 
             /* add height to the window */
             yheight[wndi] += (unsigned int)expand_by;
-            outy += (int)expand_by;
+            outy += expand_by;
             wndi++;
 
             /* move the others down */
